Update to new job deferral mechanism + misc fixes
This commit is a major rewrite and updates the binding to use the new
afb_api_queue_job() method of job deferral, instead of being
timer-based. This greatly simplifies the code and fixes the associated
timer bugs. The binding core is now also fully asynchronous.

In addition, the following changes have also been done:
- directly import the multiple API verbs using afb_api_set_verbs_v3() as
  opposed to using the controller methods.
- properly send a reply when publication startup errors out.
- use NULL when sending some replies, as opposed to wrongly putting the
  reply string in the reply itself.
- consistent use of snake case in routine and variable names.