Fixed build error seen in Koji/OBS
The Koji/OBS build environment pass the additional
-Werror=format-security option which was not present in the binding
cmake build rules. This made the build pass in native mode (where only
the binding build rules are used) but fail in Koji/OBS (due to the
differing environment).

Until both environement are fully unified, this commit adds the
problematic option to the binding build rules so as to get closer to
OBS/Koji. This allows for a better developer experience when building
across environments.

The code is also updated to fix the actual build issue.